fix: show microphone levels in capture overlays
